feat: add db iniit config settings

+++ services/bid/models.py
@@ -1,7 +1,6 @@
 from datetime import datetime
-from sqlalchemy import Column, Integer, Numeric, DateTime
-from sqlalchemy.dialects.postgresql import UUID
-from .database import Base
+from sqlalchemy import Column, Integer, DateTime
+from database import Base
 
 class Bid(Base):
     __tablename__ = "bids"
@@ -9,5 +8,5 @@ class Bid(Base):
     id         = Column(Integer, primary_key=True, autoincrement=True)
     auction_id = Column(Integer, nullable=False, index=True)
     bidder_id  = Column(Integer, nullable=False, index=True)
-    amount     = Column(Numeric(10, 2), nullable=False)
+    amount     = Column(Integer, nullable=False)
     created_at = Column(DateTime, default=datetime.utcnow, nullable=False)
